+++ packages/api/src/defaults/variants/mediaQuery.ts
@@ -1,6 +1,7 @@
 import type { MediaQuery } from "@/interfaces/variant";
 
 export const mediaQueries: MediaQuery[] = [
+	{ prefix: "pc", value: "@media (pointer: coarse)" },
 	{ prefix: "sm", value: "@media (width >= 40rem)" },
 	{ prefix: "md", value: "@media (width >= 48rem)" },
 	{ prefix: "lg", value: "@media (width >= 64rem)" },