+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 
-- Restored `README.md` file.
+- [cli] Restored `README.md` file.
 
 ## [3.8.0] - 2026-01-15